Added Sprite_Component class and updated framework for it
To add a sprite to a gameobject you only need to call the function "AddSprite("path to png")" on the gameobject. IMPORTANT: sprite must be added BEFORE setting the radius or size of the gameobject. The sprite will automatically adjust it's size/radius after any future changes to the size/radius of the gameobject.